+++ lib/volt/models/model_helpers/model_change_helpers.rb
@@ -7,8 +7,6 @@ module Volt
       base.setup_action_helpers_in_class(:before_save, :before_validate)
     end
 
-    private
-
     # Called when something in the model changes.  Saves
     # the model if there is a persistor, and changes the
     # model to not be new.
@@ -42,14 +40,16 @@ module Volt
           end
 
           # Return result inside of the validate! promise
-          result
+          result.then { self }
         end
       end
 
       # Didn't run validations
-      nil
+      self.then
     end
 
+
+    private
     # Should only be called from run_changed.  Saves the changes back to the persistor
     # and clears the tracked changes.
     def persist_changes(attribute_name)
@@ -58,8 +58,12 @@ module Volt
       # Don't save right now if we're in a nosave block
       unless Volt.in_mode?(:no_save)
         # Call the before_save callback
-        run_callbacks(:before_save)
-
+        # skip validations when running before_save, this prevents n+1, and allows
+        # the before_save to put the model into an invalid state, which you want
+        # sometimes.
+        Volt::Model.no_validate do
+          run_callbacks(:before_save)
+        end
 
         # the changed method on a persistor should return a promise that will
         # be resolved when the save is complete, or fail with a hash of errors.