+++ spec/models/model_helpers/model_helpers_spec.rb
@@ -1,0 +1,8 @@
+require 'spec_helper'
+
+describe Volt::ModelHelpers do
+  # it 'should have a root method on models that returns the root collection' do
+  #   model = store._posts.create
+  #   expect(model.root).to eq(store)
+  # end
+end